Move `StateList` to `Icinga\Module\Icingadb\Widget\ItemList`
